backend/requirements.txt
fastapi
uvicorn
python-dotenv
requests
python-multipart
pandas
scikit-learn
fairlearn
aif360
